third_party/blink/web_tests/platform/generic/fast/forms/submit-onFocus-invalidForm-expected.txt

Test that form.submit() when called on "focus" event for invalid form should submit the form

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Form is successfully submitted and content is loaded in the frame
PASS successfullyParsed is true

TEST COMPLETE

 
